記載ファイルの誤りを修正
