+++ app/channels/application_cable/connection.rb
@@ -1,8 +1,4 @@
-class ApplicationController < ActionController::Base
-  def self.render_with_signed_in_user(user, *args)
-    ActionController::Renderer::RACK_KEY_TRANSLATION['warden'] ||= 'warden'
-    proxy = Warden::Proxy.new({}, Warden::Manager.new({})).tap { |i| i.set_user(user, scope: :user) }
-    renderer = self.renderer.new('warden' => proxy)
-    renderer.render(*args)
+module ApplicationCable
+  class Connection < ActionCable::Connection::Base
   end
 end
